Move Add-Hoc commands UI code into its own plugin
