+++ src/plugins/adhoc-views/adhoc-commands.js
