+++ src/plugins/adhoc-views/templates/ad-hoc-command-form.js
