+++ src/plugins/adhoc-views/templates/ad-hoc.js
